+++ src/Cpaint/AdditionalCommands.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
+using System.Runtime.CompilerServices;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -12,7 +13,7 @@ namespace Cpaint
     static class AdditionalCommands
     {
 
-        public static void Save(IEnumerable<IFigure> figures)
+        public static async Task Save(IEnumerable<IFigure> figures)
         {
             var types = figures.Select(f => f.GetType().Name).ToArray();
             var fname = "drawing_" + Guid.NewGuid().ToString("d");
@@ -23,7 +24,7 @@ namespace Cpaint
                 item.type = types[idx++];
             }
 
-            File.WriteAllText(fname, data.ToString());
+            await File.WriteAllTextAsync(fname, data.ToString());
         }
 
     }